Set up PropertyCard content + styles
This styles the containing PropertyCard elements and starts to fill in
content, with a focus on getting the TitleImage in place.

Data from the API is now filtered to exclude properties with Available
Lot as the constructionStatus, and the logic from the instructions was
used for the getTitleImage function which also displays a placeholder
and throws a console error if no image is specified.